+++ src/commit.task.ts
@@ -1,6 +1,6 @@
 import {spawn} from '@feltjs/util/process.js';
 import {z} from 'zod';
-import {execSync} from 'child_process';
+import {execSync} from 'node:child_process';
 
 import type {Task} from './task/task';