+++ src/fs/modules.test.ts
@@ -1,6 +1,6 @@
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
-import {resolve, join} from 'path';
+import {resolve, join} from 'node:path';
 
 import {findModules, loadModules, loadModule} from './modules.js';
 import * as modTest1 from './fixtures/test1.foo.js';