+++ changelog.md
@@ -7,6 +7,8 @@
 - add `gen` plugin that defaults to watch mode
   ([#283](https://github.com/feltjs/gro/pull/283),
   [#381](https://github.com/feltjs/gro/pull/381))
+- add `--reset` and `--origin` flags to `gro deploy`
+  ([#383](https://github.com/feltjs/gro/pull/383))
 
 ## 0.79.0